+++ src/Recyclarr.Cli/Migration/RequiredMigrationException.cs
@@ -1,3 +1,0 @@
-namespace Recyclarr.Cli.Migration;
-
-internal class RequiredMigrationException() : Exception("Some REQUIRED migrations did not pass");